refactor(tools): distribute Scout command impls into concern modules
Move query commands (search, fetch, fetch_slack, research) into tools/query.rs, repository commands (repo_tree, repo_read, repo_overview) plus their candidate/readme helpers into tools/repo.rs, and ScoutBuilder into tools/builder.rs. tools.rs keeps the Scout struct, lifecycle (new/cancel_handle/github/brave), the run dispatcher, stdin resolution, and format_fetch_output.

The six dispatched methods are pub(super) because run() calls them across the module boundary; every other cross-file reference (Scout fields, lazy github()/brave(), stdin helpers) resolves via descendant access, so no field-visibility changes were needed. All six test modules stay in tools.rs.

All tools files are now <=400 lines (tools.rs 265, query.rs 223, repo.rs 325, builder.rs 181). Behavior unchanged: 449 tests (default) / 454 (js-rendering) pass, identical to the pre-split counts.
